+++ add_code/run_shiny_app_for_deployment.R
@@ -1,0 +1,7 @@
+# --- Shiny app ----------------------------------------------------------------
+
+remotes::install_github("joachim-gassen/tidycovid19",
+                        force = TRUE, upgrade = "never")
+library(tidycovid19)
+shiny_covid19_spread()
+