+++ example_code.R
@@ -1,178 +1,6 @@
 # Some example code on how to use the {tidycovid19}
 # Not part of the package of the package itself
 
-# --- Code to generate the cached data -----------------------------------------
-
-# remotes::install_github("joachim-gassen/tidycovid19")
-library(dplyr)
-library(lubridate)
-library(tidycovid19)
-library(stringr)
-
-source("scrape_apple_mtr_url.R", echo = FALSE)
-
-acaps <- download_acaps_npi_data()
-saveRDS(acaps, "cached_data/acaps_npi.RDS", version = 2)
-wblist <- download_wbank_data(var_def = TRUE)
-saveRDS(wblist, "cached_data/wbank.RDS", version = 2)
-jhu_list <- download_jhu_csse_covid19_data(
-  type = c("country", "country_region", "us_county")
-)
-saveRDS(jhu_list, "cached_data/jhu_csse_covid19.RDS", version = 2)
-
-amtr_url <- scrape_apple_mtr_url()
-amtr_list <- download_apple_mtr_data(
-  amtr_url,
-  type = c("country", "country_region", "country_city")
-)
-saveRDS(amtr_list, "cached_data/apple_mtr.RDS")
-
-gcmr_list <- download_google_cmr_data(
-  type = c("country", "country_region", "us_county")
-)
-saveRDS(gcmr_list, "cached_data/google_cmr.RDS", version = 2)
-
-gtlist <- download_google_trends_data(
-  type = c('country', 'country_day', 'region', 'city')
-)
-saveRDS(gtlist, "cached_data/google_trends.RDS", version = 2)
-
-oxlist <- download_oxford_npi_data(type = c("measures", "index"))
-saveRDS(oxlist, "cached_data/oxford_npi.RDS", version = 2)
-
-# Code from download_merged_data() to avoid reloading the data
-
-jhu_list <- readRDS("cached_data/jhu_csse_covid19.RDS")
-
-cases <- jhu_list[[1]] %>%
-    select(-timestamp)
-
-npis <- readRDS("cached_data/acaps_npi.RDS") %>%
-    mutate(npi_date = ymd(date_implemented)) %>%
-    rename(npi_type = category) %>%
-    select(iso3c, npi_date, log_type, npi_type)
-
-amtr_list <- readRDS("cached_data/apple_mtr.RDS")
-
-amtr <- amtr_list[[1]] %>%
-  select(-timestamp) %>%
-  rename_at(vars(-iso3c, -date), ~ paste0("apple_mtr_", .))
-
-gcmr_list <- readRDS("cached_data/google_cmr.RDS")
-
-gcmr <- gcmr_list[[1]] %>%
-  select(-timestamp) %>%
-  rename_at(vars(-iso3c, -date), ~ paste0("gcmr_", .))
-
-gtrends_list <- readRDS("cached_data/google_trends.RDS")
-
-gtrends_cd <- gtrends_list[[2]] %>%
-  select(-timestamp)
-
-gtrends_c <- gtrends_list[[1]] %>%
-  rename(gtrends_country_score = gtrends_score) %>%
-  select(-timestamp)
-
-wb_list <- readRDS("cached_data/wbank.RDS")
-wbank <-  wb_list[[1]] %>%
-  select(-country, -timestamp)
-
-calc_npi_measure <-function(type, var_name) {
-  my_npi <- npis %>% filter(npi_type == type)
-  cases %>%
-  left_join(
-      my_npi %>%
-        rename(date = npi_date) %>%
-        mutate(npi = ifelse(log_type == "Phase-out measure", -1, 1)) %>%
-        select(iso3c, date, npi) %>%
-        group_by(iso3c, date) %>%
-        summarise(npi = sum(npi)),
-      by = c("iso3c", "date")
-    ) %>%
-    group_by(iso3c) %>%
-    mutate(
-      npi = ifelse(is.na(npi), 0, npi),
-      sum_npi = cumsum(npi)
-    ) %>%
-    ungroup() %>%
-    select(.data$iso3c, .data$date, .data$sum_npi) -> df
-
-  names(df)[3] <- var_name
-  df
-}
-
-# 2020-04-01: There is a new populated category in the ACAPS NPI data
-#             "Humanitarian exemption". I do not code it for the time
-#             being as it contains only two Irish cases (parking for
-#             essential workers and leeway for pharamacisist)
-
-# 2020-04-16: The category "Social and economic measures" has been renamed
-#             to "Governance and socio-economic measures" in the ACAPS data.
-#             I reflect this name change by renaming the variable 'soc_econ'
-#             'gov_soc_econ'.
-
-merged <- cases %>%
-  left_join(
-    calc_npi_measure("Social distancing", "soc_dist"),
-    by = c("iso3c", "date")
-  ) %>%
-  left_join(
-    calc_npi_measure("Movement restrictions", "mov_rest"),
-    by = c("iso3c", "date")
-  ) %>%
-  left_join(
-    calc_npi_measure("Public health measures", "pub_health"),
-    by = c("iso3c", "date")
-  ) %>%
-  left_join(
-    calc_npi_measure("Governance and socio-economic measures", "gov_soc_econ"),
-    by = c("iso3c", "date")
-  ) %>%
-  left_join(
-    calc_npi_measure("Lockdown", "lockdown"),
-    by = c("iso3c", "date")
-  ) %>%
-  left_join(amtr, by = c("iso3c", "date")) %>%
-  left_join(gcmr, by = c("iso3c", "date")) %>%
-  left_join(gtrends_cd, by = c("iso3c", "date")) %>%
-  left_join(gtrends_c, by = "iso3c") %>%
-  left_join(wbank, by = "iso3c") %>%
-  group_by(iso3c) %>%
-  mutate(
-    has_npi = max(soc_dist) + max(mov_rest) +
-      max(.data$pub_health) + max(gov_soc_econ) +
-      max(lockdown) > 0,
-    soc_dist = ifelse(has_npi, soc_dist, NA),
-    mov_rest = ifelse(has_npi, mov_rest, NA),
-    pub_health = ifelse(has_npi, pub_health, NA),
-    gov_soc_econ = ifelse(has_npi, gov_soc_econ, NA),
-    lockdown = ifelse(has_npi, lockdown, NA)
-  ) %>%
-  select(-has_npi) %>%
-  ungroup() %>%
-  mutate(timestamp = Sys.time())
-
-saveRDS(merged, "cached_data/merged.RDS", version = 2)
-
-
-
-# --- Shiny app ----------------------------------------------------------------
-
-remotes::install_github("joachim-gassen/tidycovid19",
-                        force = TRUE, upgrade = "never")
-library(tidycovid19)
-shiny_covid19_spread()
-
-
-# --- Customize shiny app ------------------------------------------------------
-
-shiny_covid19_spread(plot_options = list(
-  type = "deaths", min_cases = 100, min_by_ctry_obs = 10,
-  edate_cutoff = 40, per_capita = FALSE, cumulative = FALSE, change_ave = 7,
-  highlight = c("FRA", "DEU", "ITA", "ESP", "GBR", "USA"),
-  intervention = "lockdown"
-))
-
 
 # --- Some visuals -------------------------------------------------------------
 
@@ -194,6 +22,15 @@ plot_covid19_spread(
   highlight = c("ITA", "ESP", "FRA", "DEU", "USA", "BEL", "FRA", "NLD", "GBR"),
   intervention = "lockdown"
 )
+
+# --- Customize shiny app ------------------------------------------------------
+
+shiny_covid19_spread(plot_options = list(
+  type = "deaths", min_cases = 100, min_by_ctry_obs = 10,
+  edate_cutoff = 40, per_capita = FALSE, cumulative = FALSE, change_ave = 7,
+  highlight = c("FRA", "DEU", "ITA", "ESP", "GBR", "USA"),
+  intervention = "lockdown"
+))
 
 
 # --- Example clipping code produced by shiny_covid19_spread() ------------------
@@ -363,3 +200,69 @@ ggplot(df, aes(x = date, fill = category, weight = nobs)) +
         legend.background = element_rect(fill = "white", color = NA),
         legend.title = element_text(size = 8),
         legend.text = element_text(size = 7))
+
+
+# --- Use regional data --------------------------------------------------------
+
+library(tidyverse)
+library(tidycovid19)
+df <- download_google_cmr_data(type = "country_region", cached = TRUE)
+
+df %>% filter(iso3c == "DEU") %>%
+  ggplot(aes(x = date, y = retail_recreation, color = region)) +
+  geom_line()
+
+
+# --- Plot Oxford Data ---------------------------------------------
+
+library(tidyverse)
+library(tidycovid19)
+df <- download_oxford_npi_data(type = "index", cached = TRUE)
+
+df %>% group_by(date) %>%
+  summarise(
+    mn_si = mean(stringency_index, na.rm = TRUE),
+    ci_si =  1.96*(sd(stringency_index, na.rm = TRUE)/sqrt((n() - 1)))
+  ) %>%
+  ggplot(aes(x = date, y = mn_si)) +
+  geom_line()  +
+  geom_errorbar(
+    aes(
+      ymin= mn_si - ci_si,
+      ymax = mn_si + ci_si
+    ),
+    width = 0.2
+  )
+
+df <- download_oxford_npi_data(type = "measures", cached = TRUE)
+
+df %>% filter(
+  npi_type != "Emergency investment in healthcare",
+  npi_type != "Investment in vaccines",
+  npi_measure != 0
+) %>%
+  ggplot(aes(x = date, fill = npi_type, weight = npi_measure)) +
+  geom_histogram(position = "stack", binwidth = 7)
+
+# --- Plot daily new cases as bar graph ----------------------------------------
+
+# Suggestion by AndreaPi (issue #19)
+
+library(tidyverse)
+library(tidycovid19)
+library(zoo)
+
+df <- download_merged_data(cached = TRUE)
+
+df %>%
+  filter(iso3c == "ITA") %>%
+  mutate(
+    new_cases = confirmed - lag(confirmed),
+    ave_new_cases = rollmean(new_cases, 7, na.pad=TRUE, align="right")
+  ) %>%
+  filter(!is.na(new_cases), !is.na(ave_new_cases)) %>%
+  ggplot(aes(x = date)) +
+  geom_bar(aes(y = new_cases), stat = "identity", fill = "lightblue") +
+  geom_line(aes(y = ave_new_cases), color ="red") +
+  theme_minimal()
+